+++ packages/rpc-cache-reader/Dockerfile
@@ -1,1 +1,7 @@
-FROM node:15-alpine
+# syntax=docker/dockerfile:1
+FROM node:15
+RUN apk add --no-cache python g++ make
+WORKDIR /rpc-cache-reader
+COPY . .
+RUN yarn install --production
+CMD ["node", "src/reader.js"]
