--- packages/rpc-cache-server/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "rpc-cache-server",
-  "version": "0.0.1",
-  "description": "A server to cache results of (expensive) RPC requests made to the Solana JSON RPC API",
-  "main": "src/reader.js",
-  "author": "Thomas Lisankie (Monadical)",
-  "license": "MIT",
-  "private": false,
-  "devDependencies": {
-    "@types/node": "^15.0.2",
-    "tslint": "^6.1.3",
-    "typescript": "^4.2.4"
-  },
-  "dependencies": {
-    "@solana/web3.js": "^1.10.0",
-    "axios": "^0.21.1",
-    "cors": "^2.8.5",
-    "json-rpc-2.0": "^0.2.16",
-    "redis": "^3.1.2",
-    "typescript": "^4.2.4"
-  }
-}
