--- packages/rpc-cache-server/tsconfig.json
@@ -1,32 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "esnext",
-    "target": "es2019",
-    "lib": [
-      "dom",
-      "dom.iterable",
-      "esnext"
-    ],
-    "allowJs": true,
-    "skipLibCheck": true,
-    "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true,
-    "strict": true,
-    "forceConsistentCasingInFileNames": true,
-    "noFallthroughCasesInSwitch": true,
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "downlevelIteration": true,
-    "noEmit": true,
-    "jsx": "react",
-    "typeRoots": [
-      "types",
-      "../../types",
-      "../../node_modules/@types"
-    ],
-    "moduleResolution": "node"
-  },
-  "include": [
-    "src"
-  ]
-}
