+++ packages/rpc-cache-writer/src/config.js
@@ -1,0 +1,6 @@
+const settings = {
+  programIDs: ["WormT3McKhFJ2RkiGpdw9GKvNCrB2aB54gb2uV9MfQC"],
+  cacheFunctions: ["getProgramAccounts"]
+};
+
+module.exports = settings;
